Rationalise save and return controller
Split the massive `SaveAndReturnController` into 2 separate controllers by leaving the functionality that logically belongs to the "save" process into the original controller, but shifting everything that belongs to the "return/resume" process to a separate `ResumeController`.

Additionally some common functionality has been shifted to a concern, although I'm sure a lot more could be shifted with a bit of refactor but didn't want to fundamentally change how this is done an the reasoning behind all this.

The ultimate objective of this refactor (apart from having code that is easier to understand and maintain) is to be able to introduce authorisation in one controller but not in the other controller more easily from the runner side.
