+++ app/controllers/metadata_presenter/save_and_return_controller.rb
@@ -1,25 +1,13 @@
+require_relative 'concerns/save_and_return'
+
 module MetadataPresenter
   class SaveAndReturnController < EngineController
-    helper_method :secret_questions, :page_slug, :confirmed_email, :get_service_name, :get_uuid, :label_text
+    include MetadataPresenter::SaveAndReturn
+
+    helper_method :secret_questions, :confirmed_email
 
     def show
       @saved_form = SavedForm.new
-    end
-
-    def page_slug
-      if params && params[:page_slug].present?
-        return params[:page_slug]
-      end
-      if session['returning_slug'].present?
-        return session['returning_slug']
-      end
-      if session['saved_form'].present?
-        return session['saved_form']['page_slug']
-      end
-
-      if params['saved_form'].present?
-        params['saved_form']['page_slug']
-      end
     end
 
     def confirmed_email
@@ -77,115 +65,12 @@ module MetadataPresenter
       end
     end
 
-    def return
-      response = get_saved_progress(get_uuid)
-
-      if response.status == 404
-        redirect_to '/record_error' and return
-      end
-
-      if response.status == 400
-        redirect_to '/record_failure' and return
-      end
-
-      if response.status == 422
-        redirect_to '/already_used' and return
-      end
-
-      @saved_form = SavedForm.new.from_json(response.body.to_json)
-      @resume_form = ResumeForm.new(@saved_form.secret_question)
-    end
-
-    def get_uuid
-      if params[:uuid].present?
-        return params[:uuid]
-      end
-
-      params[:resume_form][:uuid].presence
-    end
-
-    # rubocop:disable Style/RescueStandardError
-    def submit_secret_answer
-      response = get_saved_progress(get_uuid)
-
-      if response.status != 200
-        if response.status == 400
-          redirect_to '/record_failure' and return
-        end
-
-        redirect_to '/record_error' and return
-      end
-
-      @saved_form = SavedForm.new.from_json(response.body.to_json)
-      @resume_form = ResumeForm.new(@saved_form.secret_question)
-      @resume_form.secret_answer = resume_form_params[:resume_form][:secret_answer]
-      @resume_form.recorded_answer = @saved_form.secret_answer
-      @resume_form.attempts_remaining = 2 - @saved_form.attempts.to_i
-
-      if @resume_form.valid?
-        # redirect back to right place in form
-        session[:user_id] = @saved_form.user_id
-        session[:user_token] = @saved_form.user_token
-        session[:returning_slug] = @saved_form.page_slug
-
-        invalidate_record(@saved_form.id)
-
-        if @saved_form.service_version == service.version_id
-          redirect_to '/resume_progress' and return
-        else
-          redirect_to '/resume_from_start' and return
-        end
-      else
-        if @resume_form.attempts_remaining <= 0
-          begin
-            increment_record_counter(@saved_form.id)
-          rescue => e
-            Rails.logger.info(e)
-            redirect_to '/record_failure' and return
-          end
-          redirect_to '/record_failure' and return
-        end
-
-        increment_record_counter(@saved_form.id)
-
-        render :return, params: { uuid: @saved_form.id }
-      end
-    end
-    # rubocop:enable Style/RescueStandardError
-
-    def resume_progress
-      @user_data = load_user_data
-
-      @page ||= service.checkanswers_page
-
-      if @page
-        @page_answers = PageAnswers.new(@page, @answered_pages)
-
-        render template: 'metadata_presenter/save_and_return/resume_progress'
-      else
-        not_found
-      end
-    end
-
     def save_progress
       session['saved_form']['user_id'] = nil
       session['saved_form']['user_token'] = nil
       session['user_id'] = nil
       session['user_token'] = nil
     end
-
-    def answered_pages
-      TraversedPages.new(service, @user_data, @page).all
-    end
-
-    def pages_presenters
-      PageAnswersPresenter.map(
-        view: view_context,
-        pages: answered_pages,
-        answers: @user_data
-      )
-    end
-    helper_method :pages_presenters
 
     def secret_questions
       [
@@ -200,6 +85,8 @@ module MetadataPresenter
 
       secret_questions.select { |s| s.id.to_s == question.to_s }.first.name
     end
+
+    private
 
     def saved_form_params
       params.permit(
@@ -217,23 +104,6 @@ module MetadataPresenter
         :authenticity_token
       )
     end
-
-    def resume_form_params
-      params.permit(
-        { resume_form: %i[secret_answer uuid] },
-        :authenticity_token
-      )
-    end
-
-    def get_service_name
-      service.service_name
-    end
-
-    def label_text(text)
-      "<h2 class='govuk-heading-m'>#{text}</h2>"
-    end
-
-    private
 
     def service_slug
       @service_slug ||= service_slug_config
