+++ app/views/metadata_presenter/resume/resume_from_start.html.erb
