+++ app/views/metadata_presenter/resume/resume_progress.html.erb
